+++ components/common/Profile.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
 import Image from 'next/image';
 import defaultProfileImg from '@/public/assets/images/default-profile.svg';
-import { UserDataType } from '@/constants/dataType';
-import { SampleUserType } from '@/constants/sampleDataType';
-import { DEVICE_SIZE } from '@/styles/DeviceSize';
+import { UserDataType } from '@/lib/types/dataType';
+import { SampleUserType } from '@/lib/types/sampleDataType';
+import { DEVICE_SIZE } from '@/lib/styles/DeviceSize';
 
 interface Props {
   user: UserDataType | SampleUserType;
