+++ components/folder/AddLinkBar.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import LinkBar from './LinkBar';
 import { useObserver } from '@/hooks/useObserver';
-import { DEVICE_SIZE } from '@/styles/DeviceSize';
+import { DEVICE_SIZE } from '@/lib/styles/DeviceSize';
 import { useState } from 'react';
 
 function AddLinkBar() {
