+++ hooks/useGetSampleLinks.ts
@@ -1,6 +1,6 @@
 import useGetData from './useGetData';
-import { SAMPLE } from '@/constants/path';
-import { SampleFolderType, SampleLinkType } from '@/constants/sampleDataType';
+import { SAMPLE } from '@/lib/constants/path';
+import { SampleFolderType, SampleLinkType } from '@/lib/types/sampleDataType';
 
 function useGetSampleLinks(): SampleLinkType[] {
   const folderData = useGetData<SampleFolderType>(SAMPLE.folder);
