+++ lib/types/sampleDataType.ts
